# openapi.yaml
openapi: 3.0.0
info:
  title: TradeOn API
  version: 1.0.0
  description: 'TradeOn은 사용자 간 물물교환 기반의 중고 거래 플랫폼입니다.


    이 API 문서는 다음과 같은 기능들을 포함합니다:

    - 사용자 회원가입, 로그인 및 정보 조회

    - 교환 요청 및 거래 이력 관리

    - 사용자 리뷰 시스템

    - 가치 평가 게시판

    - 하이브리드 검색기능

    - 게시물 등록/삭제/조회

    - 채팅 및 거래 약속

    - 거래 데이터 분석 및 시각화


    JWT 인증이 필요한 API의 경우 Authorize 버튼을 통해 토큰을 입력해주세요.'
servers:
- url: https://673e-121-140-48-141.ngrok-free.app
  description: 로컬
...
paths:
...
  /api/v1/posts:
    post:
      tags:
        - Post
      summary: 게시글 등록
      description: 인증된 사용자가 새 게시글을 등록합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreateRequest'
      responses:
        '201':
          description: 게시글 등록 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: 게시글 등록 완료
                  post_id:
                    type: integer
                    example: 101
        '400':
          description: 필수 입력 항목 누락 또는 유효성 오류
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: 인증 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Post
      summary: 게시글 목록 조회
      description: 인증된 사용자가 카테고리 필터 및 페이지네이션으로 게시글 목록을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          description: 필터할 카테고리 (선택)
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: 페이지 번호 (빈칸시 기본값 1,최대 500 )
          required: false
          schema:
            type: integer
        - name: per_page
          in: query
          description: 페이지당 항목 수 (빈칸시 기본값 10,최대50)
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: 게시글 목록 반환
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
                  posts:
                    type: array
                    items:
                      $ref: '#/components/schemas/PostResponse'
  /api/v1/posts/{postid}:
    get:
      tags:
        - Post
      summary: 게시글 상세 조회
      description: 인증된 사용자가 특정 게시글의 상세 정보를 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 조회할 게시글 ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 게시글 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostResponse'
        '404':
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


    put:
      tags:
        - Post
      summary: 게시글 수정
      description: 작성자만 자신의 게시글을 수정할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 수정할 게시글 ID
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostUpdateRequest'
      responses:
        '200':
          description: 게시글 수정 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: 게시글 수정 완료
        '403':
          description: 수정 권한이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - Post
      summary: 게시글 삭제
      description: 작성자만 자신의 게시글을 삭제할 수 있습니다.
      security:
        - bearerAuth: []
      parameters:
        - name: postid
          in: path
          description: 삭제할 게시글 ID
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 게시글 삭제 완료
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg:
                    type: string
                    example: 게시글 삭제 완료
        '403':
          description: 삭제 권한이 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 게시글을 찾을 수 없습니다.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/posts/me:
    get:
      tags:
        - Post
      summary: 나의 게시글 목록 조회
      description: 로그인한 사용자가 작성한 게시글 목록을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: category
          in: query
          description: 필터할 카테고리 (선택)
          required: false
          schema:
            type: string
        - name: page
          in: query
          description: 페이지 번호 (빈칸시 기본값 1)
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
            maximum: 1000
        - name: per_page
          in: query
          description: 페이지당 항목 수 (빈칸시 기본값 10)
          required: false
          schema:
            type: integer
            default: 10
            maximum: 50
      responses:
        '200':
          description: 게시글 목록 조회 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  per_page:
                    type: integer
                  posts:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        title:
                          type: string
                        description:
                          type: string
                        category:
                          type: string
                        thumbnail_image_url:
                          type: string
                        user_id:
                          type: integer
                        created_at:
                          type: string
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chatrooms:
    get:
      tags:
        - Chat
      summary: 채팅방 목록 조회
      description: 인증된 사용자의 참여 중인 채팅방 목록을 최신순으로 반환합니다.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: 채팅방 목록 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Chatroom'
    post:
      tags:
        - Chat
      summary: 채팅방 생성
      description: 인증된 사용자가 새로운 채팅방을 생성합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatroomCreateRequest'
      responses:
        '201':
          description: 채팅방 생성 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chatroom'
  /api/v1/chatrooms/{chatroomId}:
    get:
      tags:
        - Chat
      summary: 채팅방 상세 조회
      description: 특정 채팅방의 상세 정보를 반환합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: chatroomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 채팅방 상세 정보 반환
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chatroom'
        '404':
          description: 채팅방 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chatrooms/{chatroomId}/messages:
    get:
      tags:
        - Chat
      summary: 메시지 목록 조회
      description: 특정 채팅방의 메시지 내역을 조회합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: chatroomId
          in: path
          required: true
          schema:
            type: integer
        - name: limit
          in: query
          description: 반환할 메시지 개수 (기본값 50)
          required: false
          schema:
            type: integer
        - name: before
          in: query
          description: 이 ISO8601 타임스탬프 이전 메시지만 조회
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: 메시지 리스트 반환
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChatMessage'
        '404':
          description: 채팅방 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags:
        - Chat
      summary: 메시지 전송
      description: 특정 채팅방에 메시지를 전송합니다.
      security:
        - bearerAuth: []
      parameters:
        - name: chatroomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatMessageSendRequest'
      responses:
        '201':
          description: 메시지 전송 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChatMessage'
        '400':
          description: 메시지 내용 누락
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 채팅방 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/chatrooms/trade-promise:
    post:
      tags:
        - Chat
      summary: 교환 약속 등록
      description: 특정 채팅방에 교환 약속을 등록합니다.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TradePromiseRequest'
      responses:
        '201':
          description: 약속 등록 완료
        '400':
          description: 필드 누락
        '404':
          description: 채팅방 없음
...
components:
  schemas:
...
    PostCreateRequest:
      type: object
      required:
        - title
        - description
        - category
      properties:
        title:
          type: string
          example: MacBook Pro 2020 판매
        description:
          type: string
          example: 거의 새 제품이며 구성품 모두 포함되어 있습니다.
        category:
          type: string
          example: 전자기기
        thumbnail_image_url:
          type: string
          example: https://example.com/macbook.jpg
    PostResponse:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
        user_id:
          type: integer
        created_at:
          type: string
          format: date-time
    PostUpdateRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        thumbnail_image_url:
          type: string
    Chatroom:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        participants:
          type: array
          items:
            type: integer
        related_post_id:
          type: integer
        last_message:
          type: string
    ChatroomCreateRequest:
      type: object
      required:
        - name
        - participant_ids
      properties:
        name:
          type: string
        participant_ids:
          type: array
          items:
            type: integer
        related_post_id:
          type: integer
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        chatroom_id:
          type: integer
        sender_id:
          type: integer
        content:
          type: string
        timestamp:
          type: string
          format: date-time
    ChatMessageSendRequest:
      type: object
      required:
        - sender_id
        - content
      properties:
        sender_id:
          type: integer
        content:
          type: string
    TradePromiseRequest:
      type: object
      required:
        - title
        - location
        - date
        - chatroom_id
      properties:
        title:
          type: string
        location:
          type: string
        date:
          type: string
          format: date-time
        chatroom_id:
          type: integer
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        msg:
          type: string
          example: 오류 메시지입니다.
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []